+++ Assets/02.Scripts/Entity/Entity.cs
@@ -41,7 +41,7 @@ public abstract class Entity : MonoBehaviour, ILocation
 
         [Range(-50, 50)]
         [Tooltip("틱마다 변경할 값 (+/-)")]
-        public int deltaPerTick = 0;
+        public float deltaPerTick = 0;
     }
     private string _locationName;
     public string locationName
@@ -178,7 +178,7 @@ public abstract class Entity : MonoBehaviour, ILocation
     [FoldoutGroup("Status Effects")] public StatusModifier sleepinessEffect = new();
     [FoldoutGroup("Status Effects")] public StatusModifier judgmentEffect = new();
 
-    public static void ApplyIfInRange(ref int actorValue, StatusModifier effect)
+    public static void ApplyIfInRange(ref float actorValue, StatusModifier effect)
     {
         if (effect == null || !effect.enabled) return;
         int min = Mathf.Clamp(effect.minValue, 0, 100);
@@ -192,7 +192,7 @@ public abstract class Entity : MonoBehaviour, ILocation
 
         if (actorValue >= min && actorValue <= max)
         {
-            int next = actorValue + effect.deltaPerTick;
+            float next = actorValue + effect.deltaPerTick;
             actorValue = Mathf.Clamp(next, 0, 100);
         }
     }
