+++ Assets/02.Scripts/Entity/Item/BodyWash.cs
@@ -11,10 +11,10 @@ public class BodyWash : Item, IUsable
     {
         if (actor != null)
         {
-            int before = actor.Cleanliness;
-            int cleanlinessIncrease = 25;
+            float before = actor.Cleanliness;
+            float cleanlinessIncrease = 25;
             actor.Cleanliness = Mathf.Min(100, actor.Cleanliness + cleanlinessIncrease);
-            int actualInc = actor.Cleanliness - before;
+            float actualInc = actor.Cleanliness - before;
             Debug.Log($"몸을 씻었습니다. 청결도 +{actualInc} ({before} → {actor.Cleanliness})");
         }
         else
